requirementsJustRTFM.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
_tflow_select=2.3.0=mkl
absl-py=1.0.0=pyhd8ed1ab_0
addict=2.4.0=pypi_0
aiohttp=3.7.0=py38h1e0a361_0
albumentations=1.0.3=pyhd8ed1ab_0
alsa-lib=1.2.3=h516909a_0
astor=0.8.1=pyh9f0ad1d_0
astunparse=1.6.3=pyhd8ed1ab_0
async-timeout=3.0.1=py_1000
attrs=21.4.0=pyhd8ed1ab_0
beautifulsoup4=4.10.0=pyha770c72_0
blas=1.0=mkl
blinker=1.4=py_1
bottleneck=1.3.2=py38heb32a55_1
brotli=1.0.9=he6710b0_2
brotlipy=0.7.0=py38h497a2fe_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f98852_0
ca-certificates=2021.10.8=ha878542_0
cachetools=4.2.4=pyhd8ed1ab_0
cairo=1.16.0=hf32fb01_1
certifi=2021.10.8=py38h578d9bd_1
cffi=1.15.0=py38h3931269_0
chardet=3.0.4=py38h924ce5b_1008
charset-normalizer=2.0.9=pyhd8ed1ab_0
click=8.0.3=py38h578d9bd_1
cloudpickle=2.0.0=pyhd3eb1b0_0
colorama=0.4.4=pyh9f0ad1d_0
cryptography=36.0.0=py38h3e25421_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pyhd3eb1b0_0
cytoolz=0.11.0=py38h7b6447c_0
dask-core=2021.10.0=pyhd3eb1b0_0
dataclasses=0.8=pyhc8e2a94_3
dbus=1.13.18=hb2f20db_0
dlib=19.22.0=py38h5ef3938_0
expat=2.4.1=h2531618_2
face_recognition=1.3.0=pyhd3deb0d_2
face_recognition_models=0.3.0=pyh9f0ad1d_0
ffmpeg=4.3.2=hca11adc_1
filelock=3.4.0=pyhd8ed1ab_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
fsspec=2021.10.1=pyhd3eb1b0_0
gast=0.4.0=pyh9f0ad1d_0
gdown=4.2.0=pyhd8ed1ab_0
geos=3.8.0=he6710b0_0
gettext=0.21.0=hf68c758_0
giflib=5.2.1=h7b6447c_0
glib=2.70.2=h780b84a_0
glib-tools=2.70.2=h780b84a_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=2.3.3=pyh6c4a22f_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
google-pasta=0.2.0=pyh8c360ce_0
graphite2=1.3.14=h23475e2_0
grpcio=1.43.0=py38hdd6454d_0
gst-plugins-base=1.18.5=hc549b11_0
gstreamer=1.18.5=ha1a6a79_0
h5py=2.10.0=nompi_py38h9915d05_106
harfbuzz=2.9.1=h83ec7ef_1
hdf5=1.10.6=hb1b8bf9_0
icu=68.1=h2531618_0
idna=3.1=pyhd3deb0d_0
iglovikov-helper-functions=0.0.53=pypi_0
imagecorruptions=1.1.2=pypi_0
imageio=2.9.0=pyhd3eb1b0_0
imgaug=0.4.0=pyhd3eb1b0_0
importlib-metadata=4.10.1=py38h578d9bd_0
imutils=0.5.4=py38h578d9bd_1
intel-openmp=2021.4.0=h06a4308_3561
jasper=1.900.1=hd497a04_4
jbig=2.1=hdba287a_0
joblib=1.1.0=pypi_0
jpeg=9d=h7f8727e_0
jpeg4py=0.1.4=pypi_0
jsonpatch=1.32=pyhd8ed1ab_0
jsonpointer=2.0=py_0
keras=2.4.3=pyhd8ed1ab_0
keras-preprocessing=1.1.2=pyhd8ed1ab_0
kiwisolver=1.3.1=py38h2531618_0
krb5=1.19.2=hcc1bbae_3
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
lerc=3.0=h295c915_0
libblas=3.9.0=12_linux64_mkl
libcblas=3.9.0=12_linux64_mkl
libclang=11.1.0=default_ha53f305_1
libdeflate=1.8=h7f8727e_5
libedit=3.1.20191231=he28a2e2_2
libevent=2.1.10=h9b69904_4
libffi=3.4.2=h295c915_2
libgcc-ng=11.2.0=h1d223b6_11
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libglib=2.70.2=h174f98d_0
libiconv=1.16=h516909a_0
libidn2=2.3.2=h7f8727e_0
liblapack=3.9.0=12_linux64_mkl
liblapacke=3.9.0=12_linux64_mkl
libllvm11=11.1.0=h3826bc1_0
libnsl=2.0.0=h7f98852_0
libogg=1.3.5=h27cfd23_1
libopencv=4.5.3=py38h5627943_1
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libpq=13.5=hd57d9b9_1
libprotobuf=3.16.0=h780b84a_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=he4da1e4_11
libtasn1=4.16.0=h27cfd23_0
libtiff=4.3.0=h6f004c6_2
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libvorbis=1.3.7=h7b6447c_0
libwebp=1.2.1=h3452ae3_0
libwebp-base=1.2.1=h7f98852_0
libxcb=1.14=h7b6447c_0
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.9.12=h72842e0_0
libzlib=1.2.11=h36c2ea0_1013
llvm-openmp=12.0.1=h4bd325d_1
locket=0.2.1=py38h06a4308_1
lz4-c=1.9.3=h295c915_1
markdown=3.3.6=pyhd8ed1ab_0
matplotlib=3.5.0=py38h578d9bd_0
matplotlib-base=3.5.0=py38h3ed280b_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
multidict=5.2.0=py38h497a2fe_1
munkres=1.1.4=py_0
mysql-common=8.0.27=ha770c72_2
mysql-libs=8.0.27=hfa10184_2
ncurses=6.2=h58526e2_4
nettle=3.7.3=hbbd107a_1
networkx=2.6.3=pyhd3eb1b0_0
nspr=4.32=h9c3ff4c_1
nss=3.73=hb5efdd6_0
numexpr=2.7.3=py38h22e1b3c_1
numpy=1.21.2=py38h20f2e39_0
numpy-base=1.21.2=py38h79a1101_0
oauthlib=3.1.1=pyhd8ed1ab_0
olefile=0.46=pyhd3eb1b0_0
opencv=4.5.3=py38h578d9bd_1
opencv-python=4.5.4.60=pypi_0
openh264=2.1.1=h780b84a_0
openssl=1.1.1l=h7f98852_0
opt_einsum=3.3.0=pyhd8ed1ab_1
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.4=pypi_0
partd=1.2.0=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pillow=8.4.0=py38h5aabda8_0
pip=21.2.4=py38h06a4308_0
pixman=0.40.0=h7f8727e_1
protobuf=3.16.0=py38h709712a_0
py-opencv=4.5.3=py38he5a9106_1
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd8ed1ab_0
pyjwt=2.3.0=pyhd8ed1ab_1
pyopenssl=21.0.0=pyhd8ed1ab_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.12.3=py38h578d9bd_8
pyqt-impl=5.12.3=py38h0ffb2e6_8
pyqt5-sip=4.19.18=py38h709712a_8
pyqtchart=5.12=py38h7400c14_8
pyqtwebengine=5.12.1=py38h7400c14_8
pysocks=1.7.1=py38h578d9bd_4
python=3.8.12=hb7a2778_2_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
python-flatbuffers=2.0=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch=1.10.0=py3.8_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pypi_0
pyu2f=0.1.5=pyhd8ed1ab_0
pywavelets=1.1.1=py38h7b6447c_2
pyyaml=6.0=py38h7f8727e_1
pyzmq=22.3.0=py38h2035c66_1
qt=5.12.9=hda022c4_4
readline=8.1=h27cfd23_0
requests=2.26.0=pyhd8ed1ab_1
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rsa=4.8=pyhd8ed1ab_0
scikit-image=0.18.3=py38h51133e4_0
scikit-learn=1.0.1=pypi_0
scipy=1.7.3=pypi_0
seaborn=0.11.0=py_0
setuptools=58.0.4=py38h06a4308_0
shapely=1.7.1=py38h1728cc4_0
six=1.16.0=pyhd3eb1b0_0
soupsieve=2.3.1=pyhd8ed1ab_0
sqlite=3.37.0=h9cd32fc_0
tensorboard=2.7.0=pyhd8ed1ab_0
tensorboard-data-server=0.6.0=py38h3e25421_1
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
tensorflow=2.4.1=mkl_py38hb2083e0_0
tensorflow-base=2.4.1=mkl_py38h43e0292_0
tensorflow-estimator=2.6.0=py38h709712a_0
termcolor=1.1.0=py_2
threadpoolctl=3.0.0=pypi_0
tifffile=2021.11.2=pypi_0
timm=0.4.12=pypi_0
tk=8.6.11=h1ccaba5_0
toolz=0.11.2=pyhd3eb1b0_0
torchfile=0.1.0=py_0
torchvision=0.11.1=py38_cu113
tornado=6.1=py38h27cfd23_0
tqdm=4.62.3=pyhd8ed1ab_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.7=pyhd8ed1ab_0
visdom=0.1.8.9=0
websocket-client=1.2.3=pyhd8ed1ab_0
werkzeug=2.0.1=pyhd8ed1ab_0
wheel=0.37.0=pyhd3eb1b0_1
wrapt=1.13.3=py38h497a2fe_1
x264=1!161.3030=h7f98852_1
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
yarl=1.7.2=py38h497a2fe_1
zeromq=4.3.4=h9c3ff4c_1
zipp=3.7.0=pyhd8ed1ab_0
zlib=1.2.11=h36c2ea0_1013
zstd=1.5.0=ha4553b6_1
